frontend: added error messages and confirmation modals for deletion/dropping. created basic CSS for visual clarity. base project functionalities complete.

+++ frontend/src/pages/Table.tsx
@@ -40,8 +40,9 @@ const Table = (): React.ReactNode => {
             table: tableName
         })
 
-        if (response.status !== "success") {
+        if (response.status !== "success"){
             console.error("Error: " + response.message);
+            alert("Error: " + response.message);
             return;
         }
 
@@ -61,23 +62,39 @@ const Table = (): React.ReactNode => {
             return;
         }
 
+        const lastColumn = columns[columns.length - 1];
+
+        if(lastColumn?.name !== column){
+            alert("Delete all columns to the right first.");
+            return;
+        }
+
+        let droppingConfirmation = confirm(`Are you sure you want to drop the column ${column}?`);
+
+        if (!droppingConfirmation)
+            return;
+
         const response = await handleOperation({
             action: "drop_column",
             table: tableName,
             name: column
         });
 
-        if (response.status !== "success")
+        if (response.status !== "success"){
             console.error("Error: " + response.message);
+            alert("Error: " + response.message);
+            return;
+        }
 
         await fetchTable();
     }
 
     const castToType = (key: string, value: string | number): string | number | boolean => {
+        console.log(typeof (value))
         if (columnsTypes[key] == Type.INT || columnsTypes[key] == Type.FLOAT)
             return Number(value);
         else if (columnsTypes[key] == Type.BOOL)
-            return value === 1;
+            return value == 1;
         else
             return value;
     }
@@ -99,8 +116,12 @@ const Table = (): React.ReactNode => {
             data: dataArray
         });
 
-        if (response.status !== "success")
+        if (response.status !== "success"){
             console.error("Error: " + response.message);
+            alert("Error: " + response.message);
+            return;
+        }
+            
 
         await fetchTable();
         setNewData({});
@@ -118,30 +139,37 @@ const Table = (): React.ReactNode => {
             id: id
         })
 
-        if (response.status !== "success")
+        if (response.status !== "success"){
             console.error("Error: " + response.message);
+            alert("Error: " + response.message);
+            return;
+        }
 
         await fetchTable();
     }
 
-    const handleUpdateClick = (id: number, data: Record<string, any>) : void => {
+    const handleUpdateClick = (id: number, data: Record<string, any>): void => {
         setEditingId(id);
-        const dataObj : Record<string, any> = {};
-        
+        const dataObj: Record<string, any> = {};
+
         var i = 0;
-        for(const col of columns){
-            dataObj[col.name] = data[i++];
+        for (const col of columns) {
+            if(col.type === Type.BOOL){
+                dataObj[col.name] = data[i++] === true ? 1 : 0;
+            } else {
+                dataObj[col.name] = data[i++];
+            }
         }
         setEditData(dataObj);
     }
 
-    const updateRow = async () : Promise<void> => {
+    const updateRow = async (): Promise<void> => {
         if (!tableName) {
             console.error("Table name not found in the URL!");
             return;
         }
 
-        if(editingId === null) {
+        if (editingId === null) {
             console.error("No row selected!");
             return;
         }
@@ -158,8 +186,11 @@ const Table = (): React.ReactNode => {
             id: editingId
         });
 
-        if (response.status !== "success")
+        if (response.status !== "success"){
             console.error("Error: " + response.message);
+            alert("Error: " + response.message);
+            return;
+        }
 
         await fetchTable();
         setEditingId(null);
@@ -181,7 +212,7 @@ const Table = (): React.ReactNode => {
     }, [columns]);
 
     const handleChange = (
-        e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>, 
+        e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
         obj: Record<string, any>,
         setter: React.Dispatch<React.SetStateAction<Record<string, any>>>): void => {
         const { name, value } = e.target;
@@ -194,91 +225,108 @@ const Table = (): React.ReactNode => {
 
     return (
         <>
+            <button onClick={() => navigate("/")}>Back to Table Navigation</button>
             <div className="main">
-                <h2>{tableName}</h2>
-                <button onClick={() => navigate("/")}>Back to Table Navigation</button>
-                <table>
-                    <thead>
-                        <tr>
-                            <th>ID</th>
-                            {columns.map((column, index) => (
-                                <th key={index}>
-                                    <div>
-                                        {column.name}
-                                        <button onClick={() => dropColumn(column.name)}>Drop</button>
-                                    </div>
-                                </th>
-                            ))}
-                            <th><ColumnInput tableName={tableName} fetchTable={fetchTable} /></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <tr>
-                            <td></td>
-                            {columns.map((column, index) => (
-                                <td key={index}>
-                                    <AdaptiveInput
-                                        name={column.name}
-                                        type={column.type}
-                                        value={newData[column.name]}
-                                        placeholder={column.name}
-                                        onChange={(e) => handleChange(e, newData, setNewData)}
-                                    />
-                                </td>
-                            ))}
-                            <td>
-                                <button onClick={() => insertRow()}>Add row</button>
-                            </td>
-                        </tr>
-                        {rows.map((row, rowIndex) => (
-                            <tr key={rowIndex}>
-                                <td>{row.id}</td>
-                                {editingId !== row.id && (
-                                    <>
-                                        {row.data.map((item, itemIndex) => (
-                                            <td key={itemIndex}>{item}</td>
-                                        ))}
-                                        <td>
-                                            <div>
-                                                <button
-                                                    onClick={() => handleUpdateClick(row.id, row.data)}
-                                                >
-                                                    Edit row
-                                                </button>
-                                                <button
-                                                    onClick={() => deleteRow(row.id)}
-                                                >
-                                                    Delete row
-                                                </button>
+                <div className="table_container">
+                    <div className="table_wrapper">
+                        <table>
+                            <thead>
+                                <tr>
+                                    <th className="table_name_h"colSpan={columns.length + 2}>
+                                        {tableName}
+                                    </th>
+                                </tr>
+                                <tr>
+                                    <th>ID</th>
+                                    {columns.map((column, index) => (
+                                        <th key={index}>
+                                            <div className="table_column_h">
+                                                {column.name} ({column.type})
+                                                <button onClick={() => dropColumn(column.name)}>Drop</button>
                                             </div>
-
+                                        </th>
+                                    ))}
+                                    <th><ColumnInput tableName={tableName} fetchTable={fetchTable} /></th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <tr>
+                                    <td>-</td>
+                                    {columns.map((column, index) => (
+                                        <td key={index}>
+                                            <AdaptiveInput
+                                                name={column.name}
+                                                type={column.type}
+                                                value={newData[column.name]}
+                                                placeholder={column.name}
+                                                onChange={(e) => handleChange(e, newData, setNewData)}
+                                            />
                                         </td>
-                                    </>
-                                )}
-                                {editingId === row.id && (
-                                    <>
-                                        {columns.map((column, index) => (
-                                            <td key={index}>
-                                                <AdaptiveInput
-                                                    name={column.name}
-                                                    type={column.type}
-                                                    value={editData[column.name]}
-                                                    placeholder={column.name}
-                                                    onChange={(e) => handleChange(e, editData, setEditData)}
-                                                />
-                                            </td>
-                                        ))}
-                                        <td>
-                                            <button onClick={() => updateRow()}>Confirm edit</button>
-                                            <button onClick={() => setEditingId(null)}>Quit editing</button>
-                                        </td>
-                                    </>
-                                )}
-                            </tr>
-                        ))}
-                    </tbody>
+                                    ))}
+                                    <td className="table_add_row_td">
+                                        <button onClick={() => insertRow()}>Add row</button>
+                                    </td>
+                                </tr>
+                                {rows.map((row, rowIndex) => (
+                                    <tr key={rowIndex}>
+                                        <td>{row.id}</td>
+                                        {editingId !== row.id && (
+                                            <>
+                                                {row.data.map((item, itemIndex) => {
+                                                    let printableItem;
+                                                    if (columns[itemIndex].type === Type.BOOL)
+                                                        printableItem = item === true ? "True" : "False";
+                                                    else printableItem = item;
 
-                </table>
+                                                    return (
+                                                        <td key={itemIndex}>{printableItem}</td>
+                                                    )
+                                                })}
+                                                <td>
+                                                    <div className="table_edit_row_td">
+                                                        <button
+                                                            onClick={() => handleUpdateClick(row.id, row.data)}
+                                                        >
+                                                            Edit
+                                                        </button>
+                                                        <button
+                                                            onClick={() => deleteRow(row.id)}
+                                                        >
+                                                            Delete
+                                                        </button>
+                                                    </div>
+
+                                                </td>
+                                            </>
+                                        )}
+                                        {editingId === row.id && (
+                                            <>
+                                                {columns.map((column, index) => (
+                                                    <td key={index}>
+                                                        <AdaptiveInput
+                                                            name={column.name}
+                                                            type={column.type}
+                                                            value={editData[column.name]}
+                                                            placeholder={column.name}
+                                                            onChange={(e) => handleChange(e, editData, setEditData)}
+                                                        />
+                                                    </td>
+                                                ))}
+                                                <td>
+                                                    <button onClick={() => updateRow()}>Confirm edit</button>
+                                                    <button onClick={() => setEditingId(null)}>Quit editing</button>
+                                                </td>
+                                            </>
+                                        )}
+                                    </tr>
+                                ))}
+                            </tbody>
+
+                        </table>
+                    </div>
+                </div>
+
+
             </div>
         </>
     )
